requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
arrow==1.3.0
attrs==24.3.0
certifi==2024.12.14
dnspython==2.7.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
loguru==0.7.3
multidict==6.1.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.10
six==1.17.0
sniffio==1.3.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzlocal==5.2
yarl==1.18.3
